feat: implement SortedUniqueFloatTuple and validation for candidate ratios in sampling configuration
